php-cs-fixer cleanup

+++ lib/PhpParser/ConstExprEvaluator.php
@@ -143,13 +143,13 @@ class ConstExprEvaluator {
             return $this->evaluateConstFetch($expr);
         }
 
-	    if ($expr instanceof Expr\ClassConstFetch) {
-		    return $this->evaluateClassConstFetch($expr);
-	    }
+        if ($expr instanceof Expr\ClassConstFetch) {
+            return $this->evaluateClassConstFetch($expr);
+        }
 
-	    if ($expr instanceof Expr\Cast) {
-		    return $this->evaluateCast($expr);
-	    }
+        if ($expr instanceof Expr\Cast) {
+            return $this->evaluateCast($expr);
+        }
 
         return ($this->fallbackEvaluator)($expr);
     }
@@ -231,76 +231,79 @@ class ConstExprEvaluator {
 
     /** @return mixed */
     private function evaluateConstFetch(Expr\ConstFetch $expr) {
-	    try {
-			$name = $expr->name->name;
-			
-		    if (defined($name)) {
-			    return constant($name);
-		    }
-		} catch(\Throwable $t) {}
+        try {
+            $name = $expr->name->name;
+
+            if (defined($name)) {
+                return constant($name);
+            }
+        } catch (\Throwable $t) {
+        }
 
         return ($this->fallbackEvaluator)($expr);
     }
 
-	/** @return mixed */
-	private function evaluateClassConstFetch(Expr\ClassConstFetch $expr) {
-		try {
-			$classname = $expr->class->name;
-			$property = $expr->name->name;
+    /** @return mixed */
+    private function evaluateClassConstFetch(Expr\ClassConstFetch $expr) {
+        try {
+            $classname = $expr->class->name;
+            $property = $expr->name->name;
 
-			if ('class' === $property) {
-				return $classname;
-			}
+            if ('class' === $property) {
+                return $classname;
+            }
 
-			if (class_exists($classname)) {
-				$class = new \ReflectionClass($classname);
-				if (array_key_exists($property, $class->getConstants())) {
-					$oReflectionConstant = $class->getReflectionConstant($property);
-					if ($oReflectionConstant->isPublic()) {
-						return $class->getConstant($property);
-					}
-				}
-			}
-		} catch(\Throwable $t) {}
+            if (class_exists($classname)) {
+                $class = new \ReflectionClass($classname);
+                if (array_key_exists($property, $class->getConstants())) {
+                    $oReflectionConstant = $class->getReflectionConstant($property);
+                    if ($oReflectionConstant->isPublic()) {
+                        return $class->getConstant($property);
+                    }
+                }
+            }
+        } catch (\Throwable $t) {
+        }
 
-		return ($this->fallbackEvaluator)($expr);
-	}
+        return ($this->fallbackEvaluator)($expr);
+    }
 
-	/** @return mixed */
-	private function evaluateCast(Expr\Cast $expr) {
-		try {
-			$subexpr = $this->evaluate($expr->expr);
-			$type = get_class($expr);
-			switch ($type) {
-				case Expr\Cast\Array_::class:
-					return (array) $subexpr;
+    /** @return mixed */
+    private function evaluateCast(Expr\Cast $expr) {
+        try {
+            $subexpr = $this->evaluate($expr->expr);
+            $type = get_class($expr);
+            switch ($type) {
+                case Expr\Cast\Array_::class:
+                    return (array) $subexpr;
 
-				case Expr\Cast\Bool_::class:
-					return (bool) $subexpr;
+                case Expr\Cast\Bool_::class:
+                    return (bool) $subexpr;
 
-				case Expr\Cast\Double::class:
-					switch ($expr->getAttribute("kind")) {
-						case Expr\Cast\Double::KIND_DOUBLE:
-							return (double) $subexpr;
+                case Expr\Cast\Double::class:
+                    switch ($expr->getAttribute("kind")) {
+                        case Expr\Cast\Double::KIND_DOUBLE:
+                            return (float) $subexpr;
 
-						case Expr\Cast\Double::KIND_FLOAT:
-						case Expr\Cast\Double::KIND_REAL:
-							return (float) $subexpr;
-					}
+                        case Expr\Cast\Double::KIND_FLOAT:
+                        case Expr\Cast\Double::KIND_REAL:
+                            return (float) $subexpr;
+                    }
 
-					break;
+                    break;
 
-				case Expr\Cast\Int_::class:
-					return (int) $subexpr;
+                case Expr\Cast\Int_::class:
+                    return (int) $subexpr;
 
-				case Expr\Cast\Object_::class:
-					return (object) $subexpr;
+                case Expr\Cast\Object_::class:
+                    return (object) $subexpr;
 
-				case Expr\Cast\String_::class:
-					return (string) $subexpr;
-			}
-		} catch(\Throwable $t) {}
+                case Expr\Cast\String_::class:
+                    return (string) $subexpr;
+            }
+        } catch (\Throwable $t) {
+        }
 
-		return ($this->fallbackEvaluator)($expr);
-	}
+        return ($this->fallbackEvaluator)($expr);
+    }
 }
